refactor: extract classic layout
